requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.1
wolframalpha==5.0.0
python-dotenv==1.0.0
pydantic==2.5.0